Only enqueue at most one scroll event per frame.
It doesn't seem to make sense to do so for forced layouts.
This causes extra CPU load, and can cause thrashing issues like
in the referenced bug.

There is no spec that requires the current behavior as far as I can
tell. The CSSOM spec calls out that it should fire "when the viewport
or element is scrolled" [1]

[1] https://drafts.csswg.org/cssom-view/#eventdef-document-scroll

Bug: 988991

Change-Id: I4e4ebcf2c5bcba8111d73ebd0b2b8a037034caaa