Implement MediaCapabilities: spatialRendering decoding query
Bug: 1020943

This CL allows for spatialRendering to be queried with
MediaCapabilities.decodingInfo() for clear and EME. Changes required to blink
and media were relatively small so they were combined into this CL. This CL
does not add the ability to render spatial audio.

This functionality is behind blink feature flag MediaCapabilitiesSpatialAudio.
Note that if the flag is disabled, querying for spatialRendering will return
supported because decodingInfo ignores extraneous parameters by design.

For the media implementation, AudioType was not extended even though
spatialRendering is strictly tied to the codec. Instead, spatialRendering was
compartmentalized for modularity and to limit changes to other codepaths that
depend on AudioType.

Change-Id: I01c9853b88f72ae8cb6291f9ca49e1d82cb220d8