Do not generate layout boxes for non-rendered foreignObject.
This change effectively re-introduces the code from [1] but only let it
affect layout tree building, not style computation, which was the reason
why the code was removed in the first place.

This also fixes the resolved style for width/height returned from
getComputedStyle() to return 'auto' for a computed value of 'auto'.
This matches the Firefox behavior.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1070137

Bug: 1027905
Change-Id: Ie324de5b0693beea3e6626a36b25f7274234d15b