Disallow recursive custom element constructions
With this CL, recursive custom element constructions are no
longer allowed. I.e. this will now only run the constructor once:
  class extends HTMLElement {
    constructor() {
      super();
      customElements.upgrade(this);
    }
  }

Previously, because the constructors are called synchronously,
the custom element state never gets changed to kCustom at [1],
so there was no easy way to just check the state at the top as
the spec ([2]) states in step 4.13.5.1. This CL looks at the
custom element construction stack, and looks for a null top
element, which is the case when the existing element
recursively calls the same constructor.

[1] https://cs.chromium.org/chromium/src/third_party/blink/renderer/core/html/custom/custom_element_definition.cc?rcl=07d3d555040a175bfdf35ccc008939258273e6bd&l=212
[2] https://html.spec.whatwg.org/multipage/custom-elements.html#upgrades

Bug: 966472
Change-Id: I76e88c0b70132eee2482c304ef9e727ae1fe8fc7